Adding functions to utils.common
