+++ pages/polars_to_pandas.py
@@ -1,8 +1,9 @@
 import streamlit as st
 from pages.pages_format import pages_format
-from utils.pandas_to_polars_functions import pd_to_pl_transform_execution_time, pandas_to_polars_transformation_cached
+from utils.pandas_to_polars_functions import pd_to_pl_transform_execution_time
 from utils.execution_times import execution_times_df, calculate_percent_diff_execution_times
 from utils.plotting_functions import plot_execution_time_bar_charts, plot_execution_time_comparison_bar_charts
+from utils.common import get_first_run_execution_times, set_first_run_execution_times, clear_cache
 
 # ---------------------------------------------------------------------
 # HOME PAGE - CONFIGURATION
@@ -28,18 +29,6 @@ datasets = [1_000, 10_000, 100_000, 1_000_000, 10_000_000]
 # ---------------------------------------------------------------------
 # CACHE FUNCTION TO PERSIST FIRST RUN EXECUTION TIMES
 # ---------------------------------------------------------------------
-def get_first_run_execution_times():
-    return None
-
-
-@st.cache_data
-def set_first_run_execution_times(df):
-    return df
-
-def clear_cache():
-    pandas_to_polars_transformation_cached.clear()
-
-
 # Clear cache on page refresh (if session state is reset)
 if st.session_state.is_first_run:
     clear_cache()
